Move pages from components
