+++ config/webpack.config.dev.js
@@ -88,6 +88,8 @@ module.exports = {
       // Support React Native Web
       // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
       'react-native': 'react-native-web',
+      Components: path.resolve(__dirname, '../src/components'),
+      Pages: path.resolve(__dirname, '../src/pages'),
     },
     plugins: [
       // Prevents users from importing files from outside of src/ (or node_modules/).
