+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './common/Header';
-import Pages from '../routes';
+import Header from 'Components/common/Header';
+import Pages from './routes';
 
 class App extends React.Component {
