+++ src/pages/hunters/HunterPage.scss
